src/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
gunicorn==21.2.0
openai==0.28.0
azure-identity==1.14.0
python-dotenv==1.0.0 # Pinned due to docker-in-docker issue: https://github.com/devcontainers/features/issues/616
environs==9.5.0